A few more edits during review
